+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - **Bug Fixes**
   - [spiral/console] Fixed the problem with commands description with signature definition.
   - [spiral/tokenizer] Fixed the problem with using named parameters in class located by a tokenizer.
+- **Other Features**
+  - [spiral/console] Added the ability to guess **option mode**, unless it is explicitly passed in the 
+    `Spiral\Console\Attribute\Option` attribute.
 
 ## 3.6.1 - 2023-02-20
 - **Bug Fixes**